requirements.txt
flask
torch
pillow
numpy
flask 
flask-cors 
torchvision 
ultralytics 
